refactor: :recycle: Refactor RealtimeWeatherReport component
