+++ src/components/compositions/index.ts
@@ -1,6 +1,6 @@
 export * from "./city/Forecast";
-export * from "./shared/CurrentWeather";
-export * from "./shared/CurrentWeatherDisplay";
+export * from "./shared/RealtimeWeatherReport";
+export * from "./shared/DetailedWeatherInformation";
 export * from "./shared/FavoriteButton";
 export * from "./shared/WeatherIcon";
 export * from "./shared/DegreesIndicator";
